+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - feature: add info whether the user has role user and admin
 - Edit customers
 - Edit Genre
+- feature: rename customer to staff
 
 ## 1.0.0